Let E-008's Chosen cross the wire: marshal its NaN Sappho Score as null
The pricing algorithm famously cannot calculate a value for E-008's Chosen
(LotNumber 6), and encoding/json cannot marshal the resulting NaN — so any
market response containing such a listing failed to encode, spotted live on
production right after deploy. SapphoScore is now a named type that
serializes NaN as JSON null (and unmarshals null back to NaN); the terminal
UI renders the anomaly as an in-theme purple 'NaN/12' instead of breaking.
The CLI already printed NaN and is unchanged in behavior.

+++ cmd/stallionussy/main.go
@@ -1214,8 +1214,8 @@ func cmdMarket(state *cliState) {
 	fmt.Println("  " + strings.Repeat("-", 80))
 
 	for _, l := range listings {
-		sapphoStr := fmt.Sprintf("%.1f", l.SapphoScore)
-		if math.IsNaN(l.SapphoScore) {
+		sapphoStr := fmt.Sprintf("%.1f", float64(l.SapphoScore))
+		if math.IsNaN(float64(l.SapphoScore)) {
 			sapphoStr = "NaN"
 		}
 		fmt.Printf("  %-30s  %8d  %7s  %s\n",
